split test files into separate class files
